requirements.txt
# JCW Cost Estimator - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Processing
numpy==1.26.4
pandas==2.1.4
pymupdf==1.23.8
openpyxl==3.1.2

# Machine Learning
scikit-learn==1.7.2
scipy==1.16.2

# API & Validation
pydantic==2.5.0

# CORS & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment Variables
python-dotenv==1.0.0

# Takeoff & Validation
pdfminer.six==20240706
jsonschema==4.23.0
pyyaml

# Layout Analysis & OCR (R2.1)
layoutparser==0.3.4
opencv-python-headless>=4.8
ocrmypdf>=15.0.0 ; optional
pytesseract>=0.3.10 ; optional
paddleocr>=2.7.0 ; optional

# Date parsing
python-dateutil==2.8.2
